Decouple skills listing from agent lifecycle
Bugbot flagged that the skills:list handler returned empty whenever getAgent()
was null — which happens during NEW_SESSION (agent stopped then restarted)
while the app stays in the ready phase — and that the panel could hang on
"Loading…" if the bridge was unavailable.

Skills are static files, so read them via pi's loadSkills() straight from disk
through a new @repo/pi-driver listSkills wrapper. The handler no longer depends
on a running agent, and the panel fetches on mount (falling back to an empty
list) instead of gating on the ready phase.
