+++ packages/pi-driver/src/index.ts
@@ -3,9 +3,10 @@ export type {
   PiAgentConfig,
   PiAgentEventListener,
   PiAgentTool,
-  PiAgentSkill,
   PiAgentStatus,
 } from "./agent";
+export { listSkills } from "./skills";
+export type { PiAgentSkill, ListSkillsOptions } from "./skills";
 export { createAuthStorage, hasAuth, loginWithProvider, type LoginCallbacks } from "./auth";
 export { resolveModel } from "./model";
